+++ core/src/lib/sat/boolean_expression.rs
@@ -1,64 +1,4 @@
-#[derive(Clone, Copy)]
-pub struct Literal {
-    pub variable: i32,
-    pub negate: bool,
-}
-
-impl Literal {
-    pub fn dump(&self) -> i32 {
-        match self.negate {
-            true => -self.variable,
-            false => self.variable,
-        }
-    }
-}
-
-pub struct Clause {
-    pub literals: Vec<Literal>,
-}
-
-impl Clause {
-    pub fn dump(&self) -> Vec<i32> {
-        let lits = self.literals.iter().map(|l| l.dump());
-        Vec::from_iter(lits)
-    }
-
-    pub fn add(&self, other: &Self) -> Clause {
-        let mut literals = self.literals.clone();
-        literals.extend(other.literals.iter());
-
-        Clause {
-            literals,
-        }
-    }
-}
-
-pub struct CnfBuilder {
-    pub num_variables: i32,
-    pub clauses: Vec<Clause>,
-}
-
-impl Default for CnfBuilder {
-    fn default() -> Self {
-        Self { num_variables: 0, clauses: Vec::new() }
-    }
-}
-
-impl CnfBuilder {
-    pub fn add_variable(&mut self) -> i32 {
-        self.num_variables += 1;
-        self.num_variables
-    }
-
-    pub fn dump(&self) -> Vec<Vec<i32>> {
-        Vec::from_iter(
-            self.clauses
-                .as_slice()
-                .iter()
-                .map(|c| c.dump())
-        )
-    }
-}
+use super::{ Clause, Literal };
 
 #[derive(Clone)]
 pub enum BooleanExpressionMode {
@@ -68,9 +8,9 @@ pub enum BooleanExpressionMode {
 
 #[derive(Clone)]
 pub struct BooleanExpression {
-    mode: BooleanExpressionMode,
-    literals: Vec<Literal>,
-    expressions: Vec<BooleanExpression>,
+    pub mode: BooleanExpressionMode,
+    pub literals: Vec<Literal>,
+    pub expressions: Vec<BooleanExpression>,
 }
 
 impl Default for BooleanExpression {
@@ -187,38 +127,6 @@ fn distribute(left: Vec<Clause>, right: Vec<Clause>) -> Vec<Clause> {
     }
 
     result
-}
-
-// x1 | x2
-pub fn Or(x1: BooleanExpression, x2: BooleanExpression) -> BooleanExpression {
-    BooleanExpression {
-        mode: BooleanExpressionMode::Or,
-        expressions: vec![x1, x2],
-        ..Default::default()
-    }
-}
-
-// x1 & x2
-pub fn And(x1: BooleanExpression, x2: BooleanExpression) -> BooleanExpression {
-    BooleanExpression {
-        mode: BooleanExpressionMode::Or,
-        expressions: vec![x1, x2],
-        ..Default::default()
-    }
-}
-
-pub fn Neg(mut x1: BooleanExpression) -> BooleanExpression {
-    x1.negate();
-    x1
-}
-
-// x1 => x2
-pub fn Implies(x1: BooleanExpression, x2: BooleanExpression) -> BooleanExpression {
-    BooleanExpression {
-        mode: BooleanExpressionMode::Or,
-        expressions: vec![Neg(x1), x2],
-        ..Default::default()
-    }
 }
 
 #[cfg(test)]
